+++ src/main/java/io/github/astrapi69/checksum/package-info.java
@@ -1,0 +1,5 @@
+/**
+ * This package provides helper classes for calculate checksum from byte arrays, files, string and
+ * objects
+ */
+package io.github.astrapi69.checksum;