+++ src/test/java/io/github/astrapi69/checksum/DirectoryChecksumTest.java
@@ -1,0 +1,5 @@
+package io.github.astrapi69.checksum;
+
+public class DirectoryChecksumTest
+{
+}