+++ src/test/java/io/github/astrapi69/checksum/ObjectChecksumExtensionsTest.java
@@ -46,7 +46,7 @@ public class ObjectChecksumExtensionsTest extends AbstractTestCase<Long, Long>
 
 
 	/**
-	 * Test method for {@link ObjectChecksumExtensions#getChecksums(String, Serializable[])}
+	 * Test method for {@link ObjectChecksumExtensions#getChecksum(String, Serializable[])}
 	 *
 	 * @throws NoSuchAlgorithmException
 	 *             is thrown if instantiation of the SecretKeyFactory object fails.
@@ -66,7 +66,7 @@ public class ObjectChecksumExtensionsTest extends AbstractTestCase<Long, Long>
 
 		anotherPerson = Person.builder().gender(Gender.FEMALE).name("Anton").married(false)
 			.about("I'm a dev guy").nickname("admin").build();
-		actual = ObjectChecksumExtensions.getChecksums(MdAlgorithm.MD2.getAlgorithm(), person,
+		actual = ObjectChecksumExtensions.getChecksum(MdAlgorithm.MD2.getAlgorithm(), person,
 			anotherPerson);
 		expected = "8cb07185f0ca40fb0294514fe88ad3cd";
 		assertEquals(expected, actual);
@@ -237,6 +237,7 @@ public class ObjectChecksumExtensionsTest extends AbstractTestCase<Long, Long>
 
 		expected = "b1def362c48394716231bb7e42c14b3c";
 		actual = ObjectChecksumExtensions.getChecksum(person, MdAlgorithm.MD2.getAlgorithm());
+		actual = ObjectChecksumExtensions.getChecksum(MdAlgorithm.MD2.getAlgorithm(), person);
 		assertEquals(expected, actual);
 		actualLength = actual.length();
 		expectedLength = 32;
@@ -275,233 +276,6 @@ public class ObjectChecksumExtensionsTest extends AbstractTestCase<Long, Long>
 		actualLength = actual.length();
 		expectedLength = 128;
 		assertEquals(expectedLength, actualLength);
-	}
-
-	/**
-	 * Test method for {@link ObjectChecksumExtensions#getChecksum(String, Algorithm)}
-	 *
-	 * @throws NoSuchAlgorithmException
-	 *             is thrown if instantiation of the SecretKeyFactory object fails.
-	 */
-	@Test
-	public void testGetChecksumStringAlgorithm() throws NoSuchAlgorithmException
-	{
-		String expected;
-		String actual;
-		String text;
-		int expectedLength;
-		int actualLength;
-
-		text = "Lorem ipsum dolor sit amet, sea consul verterem perfecto id. Alii prompta electram te nec, at minimum copiosae quo. Eos iudico nominati oportere ei, usu at dicta legendos. In nostrum insolens disputando pro, iusto equidem ius id.";
-
-		expected = "2016860f2f742ecce65a6c246922cdba";
-		actual = ObjectChecksumExtensions.getChecksum(text, MdAlgorithm.MD2);
-		assertEquals(expected, actual);
-		actualLength = actual.length();
-		expectedLength = 32;
-		assertEquals(expectedLength, actualLength);
-
-		expected = "56a3c95014fb72a9adbeb7f16c474bba";
-		actual = ObjectChecksumExtensions.getChecksum(text, MdAlgorithm.MD5);
-		assertEquals(expected, actual);
-		actualLength = actual.length();
-		assertEquals(expectedLength, actualLength);
-
-		expected = "e41a81525dbfb95f254e0558c8d15969b661ccc0";
-		actual = ObjectChecksumExtensions.getChecksum(text, HashAlgorithm.SHA_1);
-		assertEquals(expected, actual);
-		actualLength = actual.length();
-		expectedLength = 40;
-		assertEquals(expectedLength, actualLength);
-
-		expected = "b930e875225e5ec0291e1ccc39ab3ddf9e5de10b54f48d8470620174eea5491c";
-		actual = ObjectChecksumExtensions.getChecksum(text, HashAlgorithm.SHA_256);
-		assertEquals(expected, actual);
-		actualLength = actual.length();
-		expectedLength = 64;
-		assertEquals(expectedLength, actualLength);
-
-		expected = "1f9d7250a02aa6ad4f61425921430714e333185b4555e6c15b7911cbf096c16e649bf2891d3003ef51d65644423b3dfd";
-		actual = ObjectChecksumExtensions.getChecksum(text, HashAlgorithm.SHA_384);
-		assertEquals(expected, actual);
-		actualLength = actual.length();
-		expectedLength = 96;
-		assertEquals(expectedLength, actualLength);
-
-		expected = "4ce3fd741dc5459beda2c7ffed7e7561667e84612651935f2cf50e973f03b3094ad83e225f0ce340b688929d5fee9d6785a362a9af5b12350b4e85b9e8d3a8c4";
-		actual = ObjectChecksumExtensions.getChecksum(text, HashAlgorithm.SHA_512);
-		assertEquals(expected, actual);
-		actualLength = actual.length();
-		expectedLength = 128;
-		assertEquals(expectedLength, actualLength);
-
-		text = "secret Message";
-
-		expected = "5cc16e663491726545c13ec2012f4601";
-		actual = ObjectChecksumExtensions.getChecksum(text, MdAlgorithm.MD2);
-		assertEquals(expected, actual);
-		actualLength = actual.length();
-		expectedLength = 32;
-		assertEquals(expectedLength, actualLength);
-
-		expected = "25659bd9db98ecc3c2077d44e69607b8";
-		actual = ObjectChecksumExtensions.getChecksum(text, MdAlgorithm.MD5);
-		assertEquals(expected, actual);
-		actualLength = actual.length();
-		assertEquals(expectedLength, actualLength);
-
-		expected = "874026e54b67d4f9aaf87cb14a683fb51de6f9cb";
-		actual = ObjectChecksumExtensions.getChecksum(text, HashAlgorithm.SHA_1);
-		assertEquals(expected, actual);
-		actualLength = actual.length();
-		expectedLength = 40;
-		assertEquals(expectedLength, actualLength);
-
-		expected = "8a3b3c92a8b0eb00da917c23201a9407ef7963373464076aec4c54c066e8b7aa";
-		actual = ObjectChecksumExtensions.getChecksum(text, HashAlgorithm.SHA_256);
-		assertEquals(expected, actual);
-		actualLength = actual.length();
-		expectedLength = 64;
-		assertEquals(expectedLength, actualLength);
-
-		expected = "b58a362687ab42b9bf0d8af0b4860ed262d1fd128e16ab0082723e7785a862cd129b03577312452cc24aecdb36d5406d";
-		actual = ObjectChecksumExtensions.getChecksum(text, HashAlgorithm.SHA_384);
-		assertEquals(expected, actual);
-		actualLength = actual.length();
-		expectedLength = 96;
-		assertEquals(expectedLength, actualLength);
-
-		expected = "ab29b34a26547ca4ce517d776885a5642929d9ed571a990fc764f7d0b854d6546276ca9aa45b3d88db3dc3dbf3c2f2152017d3e3e054ed6cd7a38a1f7925a746";
-		actual = ObjectChecksumExtensions.getChecksum(text, HashAlgorithm.SHA_512);
-		assertEquals(expected, actual);
-		actualLength = actual.length();
-		expectedLength = 128;
-		assertEquals(expectedLength, actualLength);
-
-	}
-
-	/**
-	 * Test method for {@link ObjectChecksumExtensions#getChecksum(String, String)}
-	 *
-	 * @throws NoSuchAlgorithmException
-	 *             is thrown if instantiation of the SecretKeyFactory object fails.
-	 */
-	@Test
-	public void testGetChecksumStringAlgorithmAsString() throws NoSuchAlgorithmException
-	{
-		String expected;
-		String actual;
-		String text;
-		int expectedLength;
-		int actualLength;
-
-		text = "Lorem ipsum dolor sit amet, sea consul verterem perfecto id. Alii prompta electram te nec, at minimum copiosae quo. Eos iudico nominati oportere ei, usu at dicta legendos. In nostrum insolens disputando pro, iusto equidem ius id.";
-
-		expected = "2016860f2f742ecce65a6c246922cdba";
-		actual = ObjectChecksumExtensions.getChecksum(text, MdAlgorithm.MD2.getAlgorithm());
-		assertEquals(expected, actual);
-		actualLength = actual.length();
-		expectedLength = 32;
-		assertEquals(expectedLength, actualLength);
-
-		expected = "56a3c95014fb72a9adbeb7f16c474bba";
-		actual = ObjectChecksumExtensions.getChecksum(text, MdAlgorithm.MD5.getAlgorithm());
-		assertEquals(expected, actual);
-		actualLength = actual.length();
-		assertEquals(expectedLength, actualLength);
-
-		expected = "e41a81525dbfb95f254e0558c8d15969b661ccc0";
-		actual = ObjectChecksumExtensions.getChecksum(text, HashAlgorithm.SHA_1.getAlgorithm());
-		assertEquals(expected, actual);
-		actualLength = actual.length();
-		expectedLength = 40;
-		assertEquals(expectedLength, actualLength);
-
-		expected = "b930e875225e5ec0291e1ccc39ab3ddf9e5de10b54f48d8470620174eea5491c";
-		actual = ObjectChecksumExtensions.getChecksum(text, HashAlgorithm.SHA_256.getAlgorithm());
-		assertEquals(expected, actual);
-		actualLength = actual.length();
-		expectedLength = 64;
-		assertEquals(expectedLength, actualLength);
-
-		expected = "1f9d7250a02aa6ad4f61425921430714e333185b4555e6c15b7911cbf096c16e649bf2891d3003ef51d65644423b3dfd";
-		actual = ObjectChecksumExtensions.getChecksum(text, HashAlgorithm.SHA_384.getAlgorithm());
-		assertEquals(expected, actual);
-		actualLength = actual.length();
-		expectedLength = 96;
-		assertEquals(expectedLength, actualLength);
-
-		expected = "4ce3fd741dc5459beda2c7ffed7e7561667e84612651935f2cf50e973f03b3094ad83e225f0ce340b688929d5fee9d6785a362a9af5b12350b4e85b9e8d3a8c4";
-		actual = ObjectChecksumExtensions.getChecksum(text, HashAlgorithm.SHA_512.getAlgorithm());
-		assertEquals(expected, actual);
-		actualLength = actual.length();
-		expectedLength = 128;
-		assertEquals(expectedLength, actualLength);
-
-		text = "secret Message";
-
-		expected = "5cc16e663491726545c13ec2012f4601";
-		actual = ObjectChecksumExtensions.getChecksum(text, MdAlgorithm.MD2.getAlgorithm());
-		assertEquals(expected, actual);
-		actualLength = actual.length();
-		expectedLength = 32;
-		assertEquals(expectedLength, actualLength);
-
-		expected = "25659bd9db98ecc3c2077d44e69607b8";
-		actual = ObjectChecksumExtensions.getChecksum(text, MdAlgorithm.MD5.getAlgorithm());
-		assertEquals(expected, actual);
-		actualLength = actual.length();
-		assertEquals(expectedLength, actualLength);
-
-		expected = "874026e54b67d4f9aaf87cb14a683fb51de6f9cb";
-		actual = ObjectChecksumExtensions.getChecksum(text, HashAlgorithm.SHA_1.getAlgorithm());
-		assertEquals(expected, actual);
-		actualLength = actual.length();
-		expectedLength = 40;
-		assertEquals(expectedLength, actualLength);
-
-		expected = "8a3b3c92a8b0eb00da917c23201a9407ef7963373464076aec4c54c066e8b7aa";
-		actual = ObjectChecksumExtensions.getChecksum(text, HashAlgorithm.SHA_256.getAlgorithm());
-		assertEquals(expected, actual);
-		actualLength = actual.length();
-		expectedLength = 64;
-		assertEquals(expectedLength, actualLength);
-
-		expected = "b58a362687ab42b9bf0d8af0b4860ed262d1fd128e16ab0082723e7785a862cd129b03577312452cc24aecdb36d5406d";
-		actual = ObjectChecksumExtensions.getChecksum(text, HashAlgorithm.SHA_384.getAlgorithm());
-		assertEquals(expected, actual);
-		actualLength = actual.length();
-		expectedLength = 96;
-		assertEquals(expectedLength, actualLength);
-
-		expected = "ab29b34a26547ca4ce517d776885a5642929d9ed571a990fc764f7d0b854d6546276ca9aa45b3d88db3dc3dbf3c2f2152017d3e3e054ed6cd7a38a1f7925a746";
-		actual = ObjectChecksumExtensions.getChecksum(text, HashAlgorithm.SHA_512.getAlgorithm());
-		assertEquals(expected, actual);
-		actualLength = actual.length();
-		expectedLength = 128;
-		assertEquals(expectedLength, actualLength);
-
-	}
-
-	/**
-	 * Test method for {@link ObjectChecksumExtensions#getChecksum(String, boolean)}
-	 */
-	@Test
-	public void testGetChecksumStringBoolean()
-	{
-		long expected;
-		long actual;
-		String text;
-
-		text = "Lorem ipsum dolor sit amet, sea consul verterem perfecto id. Alii prompta electram te nec, at minimum copiosae quo. Eos iudico nominati oportere ei, usu at dicta legendos. In nostrum insolens disputando pro, iusto equidem ius id.";
-
-		expected = 1715371676L;
-		actual = ObjectChecksumExtensions.getChecksum(text, true);
-		assertEquals(expected, actual);
-
-		expected = 2358596523L;
-		actual = ObjectChecksumExtensions.getChecksum(text, false);
-		assertEquals(expected, actual);
 	}
 
 	/**